requirement.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==4.2.1
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
cryptography==44.0.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.5
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Mako==1.3.6
MarkupSafe==3.0.2
packaging==24.2
passlib==1.7.4
pathlib==1.0.1
pluggy==1.5.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
pytest==8.3.4
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.19
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlmodel==0.0.22
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.1
